requirements.txt
supabase==2.6.0
requests
python-dotenv
